Update .gitignore to include results and experiments_results directories; refactor logging in SimulationController and enhance blockage handling in Orchestrator for improved event processing.

+++ src/main/java/com/example/plgsystem/model/Blockage.java
@@ -44,7 +44,7 @@ public class Blockage implements Serializable {
 
     public static String serializePositions(List<Position> positions) {
         return positions.stream()
-                .map(pos -> pos.getX() + "," + pos.getY())
+                .map(pos -> (int) pos.getX() + "," + (int) pos.getY())
                 .collect(Collectors.joining(","));
     }
